requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.9
django-jazzmin==3.0.1
django-modeltranslation==0.19.13
pillow==10.4.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing-extensions==4.13.2
